chenge to bigint

+++ db/migrate/20191020083635_create_touring_members.rb
@@ -1,8 +1,8 @@
 class CreateTouringMembers < ActiveRecord::Migration[5.2]
   def change
     create_table :touring_members do |t|
-      t.integer :user_id
-      t.integer :team_id
+      t.bigint :user_id
+      t.bigint :team_id
       t.timestamps
     end
   end
